Fix candidates caching
If there are matches found, they are cached for a case when journal
contains an error and ledger is unable to get candidates. This is wrong
as on the next match attempt and invalid journal file stale (filtered)
cache is used.

This commit fixes this scenario by caching all candidates returned from
ledger command (when successful) and then collects matches from the
cache. In case of invalid journal file matches will be still provided
using previously cached, unfiltered values.